backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
aiofiles
pillow
imageio
httpx
python-dotenv
cloudinary
python-jose[cryptography]
requests
